# apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "URL Shortener API",
    "description": "A simple API to shorten URLs.",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "URLShortenerService"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/original/{shortCode}": {
      "get": {
        "summary": "Retrieves the original URL for a given short code.",
        "operationId": "URLShortenerService_GetOriginalURL",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetOriginalURLResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "shortCode",
            "description": "The short code to look up.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "URLShortenerService"
        ]
      }
    },
    "/api/v1/shorten": {
      "post": {
        "summary": "Creates a short code for a given URL.",
        "operationId": "URLShortenerService_ShortenURL",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ShortenURLResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ShortenURLRequest"
            }
          }
        ],
        "tags": [
          "URLShortenerService"
        ]
      }
    },
    "/{shortCode}": {
      "get": {
        "summary": "Redirects a short URL to the original URL",
        "description": "Given a short code, this endpoint performs an HTTP 302 redirect to the corresponding original URL.\n\n**Note for Swagger UI users:** Due to browser security restrictions, clicking 'Execute' will not perform a real redirect. Instead, it will return a `200 OK` response with the destination URL in the `Location` header, allowing you to see what the redirect target would be.",
        "operationId": "URLShortenerService_Redirect",
        "responses": {
          "200": {
            "description": "OK. The original URL is returned in the 'Location' header.",
            "schema": {
              "$ref": "#/definitions/v1RedirectResponse"
            },
            "headers": {
              "Location": {
                "description": "The original URL to redirect to.",
                "type": "string"
              }
            }
          },
          "302": {
            "description": "Found. Redirects to the original URL.",
            "schema": {},
            "headers": {
              "Location": {
                "description": "The original URL to redirect to.",
                "type": "string"
              }
            }
          },
          "404": {
            "description": "Not Found. The short code does not exist.",
            "schema": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "shortCode",
            "description": "The short code to redirect to the original URL.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Redirect"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1GetOriginalURLResponse": {
      "type": "object",
      "properties": {
        "originalUrl": {
          "type": "string",
          "description": "The original long URL."
        }
      }
    },
    "v1RedirectResponse": {
      "type": "object"
    },
    "v1ShortenURLRequest": {
      "type": "object",
      "properties": {
        "originalUrl": {
          "type": "string",
          "description": "The original URL to shorten. Must be a valid, absolute URL."
        }
      }
    },
    "v1ShortenURLResponse": {
      "type": "object",
      "properties": {
        "shortCode": {
          "type": "string",
          "description": "The generated short code."
        }
      }
    }
  }
}